refactoring strukture
