+++ backend-kb/routes/asistants/projectsRoute.js
@@ -57,31 +57,4 @@ projectsRouter.post('/setdeadline', (req, res) => {
     }
 });
 
-projectsRouter.post('/bodovanjeGrupe', (req, res) => {
-    res.setHeader('Content-Type', 'application/json');
-
-    projectUtils.provjeraParametaraBodovanjeProjektneGrupe(req.body, (cb) => {
-        if(cb.ispravno) {
-            projectUtils.upisBodovaProjektneGrupe(req.body, (err) => {
-                if(err) {
-                    res.send(JSON.stringify({
-                        message: 'Doslo je do greske sa bazom.'
-                    }));
-                }
-                else {
-                    res.send(JSON.stringify({
-                        message: 'Uspjesno bodovan projekat.'
-                    }));
-                }
-            })
-        }
-        else
-        {
-            res.send(JSON.stringify({
-                message: cb.poruka
-            }));
-        }
-    });
-});
-
 module.exports = projectsRouter;
